# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted electronics parts inventory. Track quantity, package, brand, location, and vendor pricing across your parts, scan a distributor barcode to enrich a part, and generate bin and part labels. All protected endpoints require a Bearer JWT (or `fbin_pat_*` personal access token) in the Authorization header.",
        "title": "FireBin API",
        "contact": {
            "name": "FireBin",
            "url": "https://github.com/FireBall1725/firebin"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        },
        "version": "26.7.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/boards/{id}/pick-list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Compute what to pull from stock to build N of a board.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Board pick list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "number of boards to build",
                        "name": "quantity",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PickList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return recent background tasks, newest first, optionally filtered.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "List tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by task type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "maximum tasks to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Task"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete every finished task and its logs; running or queued tasks are kept.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Clear finished tasks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int64"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return one task with its progress and result.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.Task"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/cancel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Request cancellation of a running or queued task.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Cancel task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the task's log lines after the given after_id cursor.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "return logs after this id",
                        "name": "after_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.JobLog"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/retry": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enqueue a fresh job from a finished task's arguments.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Retry task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_firelabsca_firebin-api_internal_models.JobLog": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "task_id": {
                    "type": "string"
                },
                "ts": {
                    "type": "string"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.PickEntry": {
            "type": "object",
            "properties": {
                "location_id": {
                    "type": "string"
                },
                "location_name": {
                    "description": "\"\" when the lot has no bin",
                    "type": "string"
                },
                "part_id": {
                    "type": "string"
                },
                "part_name": {
                    "type": "string"
                },
                "quantity": {
                    "type": "number"
                },
                "stock_item_id": {
                    "type": "string"
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.PickList": {
            "type": "object",
            "properties": {
                "board_id": {
                    "type": "string"
                },
                "board_name": {
                    "type": "string"
                },
                "copies": {
                    "description": "panel N-up (per board)",
                    "type": "integer"
                },
                "entries": {
                    "description": "sorted by location, then part",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PickEntry"
                    }
                },
                "quantity": {
                    "description": "boards to build",
                    "type": "integer"
                },
                "shortfalls": {
                    "description": "parts short of stock",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PickShortfall"
                    }
                },
                "total_units": {
                    "type": "number"
                },
                "unmatched": {
                    "description": "BOM lines with no matched part",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_firelabsca_firebin-api_internal_models.PickUnmatched"
                    }
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.PickShortfall": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number"
                },
                "part_id": {
                    "type": "string"
                },
                "part_name": {
                    "type": "string"
                },
                "required": {
                    "type": "number"
                },
                "short": {
                    "type": "number"
                }
            }
        },
        "github_com_firelabsca_firebin-api_internal_models.PickUnmatched": {
            "type": "object",
            "properties": {
                "quantity": {
                    "description": "total needed across the whole build",
                    "type": "integer"
                },
                "refs": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_firelabsca_firebin-api_internal_models.Task": {
            "type": "object",
            "properties": {
                "args_summary": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "attempt": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "max_attempts": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                },
                "progress_done": {
                    "type": "integer"
                },
                "progress_total": {
                    "type": "integer"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}